fix: 명시적 float64 변환으로 PyArrow 정수 오버플로우 방지
근본 원인:
- IEEE 754 double의 안전한 정수 범위: ±2^53 (약 ±9조)
- enterpriseValue 값: 9,589,977,647,903,276 (약 9.6조) → 초과!
- PyArrow가 object 타입을 정수로 변환 시도 → ArrowInvalid 에러

이전 시도의 문제:
- pd.to_numeric(errors='coerce')만으로는 부족
- pandas가 자동으로 타입 선택 (int64 또는 float64)
- PyArrow가 다시 정수로 변환 시도 가능

해결:
명시적으로 float64로 변환

Before:
  converted = pd.to_numeric(json_data[col], errors='coerce')
  if not converted.isna().all():
      json_data[col] = converted  # 타입 불명확

After:
  converted = pd.to_numeric(json_data[col], errors='coerce')
  if not converted.isna().all():
      json_data[col] = converted.astype('float64')  # 명시적 float64

효과:
- 큰 정수 값도 float64로 안전하게 저장
- PyArrow가 정수로 변환 시도하지 않음
- IEEE 754 범위 초과 에러 해결

+++ Quant-refactoring/data_collector/fmp_fetch_worker.py
@@ -179,7 +179,7 @@ def __fmp_worker(
         # Convert all object columns to numeric where possible
         # This handles large financial values (marketCap, enterpriseValue, totalAssets, etc.)
         # that can exceed IEEE 754 double precision safe integer range (±9 quadrillion)
-        # Use errors='coerce' to handle mixed types and None values
+        # IMPORTANT: Must explicitly convert to float64 to avoid PyArrow integer overflow
         for col in json_data.select_dtypes(include=['object']).columns:
             # Try to convert to numeric, coercing errors to NaN
             converted = pd.to_numeric(json_data[col], errors='coerce')
@@ -187,7 +187,9 @@ def __fmp_worker(
             # Only keep conversion if at least some values were successfully converted
             # This prevents pure string columns (like 'symbol', 'period') from becoming all NaN
             if not converted.isna().all():
-                json_data[col] = converted
+                # Explicitly convert to float64 to handle large integers safely
+                # PyArrow will fail if it tries to convert large integers as int64
+                json_data[col] = converted.astype('float64')
 
         # Save to Parquet file
         json_data.to_parquet(f"{file_path}/{safe_symbol+file_postfix}.parquet", index=False)
